requirements.txt
pyocr==0.5.3
Pillow==5.3.0
python_magic==0.4.15
